requirements.txt
alembic==1.16.5
anyio==4.11.0
asteval==1.0.6
astropy==7.1.0
astropy-iers-data==0.2025.9.22.0.37.25
astroquery==0.4.11
bcrypt==4.3.0
beautifulsoup4==4.13.5
bidict==0.23.1
blinker==1.9.0
blosc2==3.8.0
caugetch==0.0.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
clipboard==0.0.4
colorama==0.4.6
configparser==7.2.0
contourpy==1.3.3
cryptography==46.0.1
cycler==0.12.1
dill==0.4.0
Flask==3.1.2
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.7.1
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
fonttools==4.60.0
getpass4==0.0.14.1
h11==0.16.0
h5py==3.14.0
habanero==2.3.0
hitran-api==1.2.2.4
hjson==3.1.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
Jinja2==3.1.6
joblib==1.5.2
json-tricks==3.17.3
keyring==25.6.0
kiwisolver==1.4.9
llvmlite==0.45.0
lmfit==1.3.4
load-dotenv==0.1.0
lxml==6.0.2
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.6
more-itertools==10.8.0
mpldatacursor==0.7.1
msgpack==1.1.1
narwhals==2.5.0
ndindex==1.10.0
numba==0.62.0
numexpr==2.13.0
numpy==2.3.3
packaging==25.0
pandas==2.3.2
PeakUtils==1.3.5
periodictable==2.0.2
pillow==11.3.0
platformdirs==4.4.0
plotly==6.3.0
psutil==7.1.0
publib==0.4.0
py-cpuinfo==9.0.0
pycparser==2.23
pyerfa==2.0.1.5
PyJWT==2.10.1
pyparsing==3.2.5
pyperclip==1.10.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.2
python-socketio==5.13.0
pytz==2025.2
pyvo==1.7
PyYAML==6.0.2
radis==0.16.3
redis==6.4.0
requests==2.32.5
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.14
scipy==1.16.2
seaborn==0.13.2
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
tables==3.10.2
termcolor==3.1.0
toml==0.10.2
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
uncertainties==3.2.3
urllib3==2.5.0
vulkan==1.3.275.1
webencodings==0.5.1
Werkzeug==3.1.3
wsproto==1.2.0
